Update 10-02-19
Added a graph over the game to work like Dijsktra's algorithm for route finding, given a start position and end position, it will find the shortest route between them using the nodes in the graph. Now need to make the To and From fields, and create Path options available in game to the player, and then map the Followers path on the floor.